+++ data/sounds.lua
@@ -2,41 +2,39 @@
 -- Sound IDs
 --=====================================================================================
 
--- Retail
-muteSoundIDs = {
-    569143,  -- Achievement
-    1489546, -- Honor
-    569593,  -- Level Up
-    642841,  -- Battle Pet Level
-    4745441, -- Renown
-    568016,  -- Reputation (8473)
-    567400,  -- Quest Accepted
-    567439,  -- Quest Turned In
-    2066672  -- Trade Post
-}
-
--- Cataclysm
-muteSoundIDs_c = {
-    569143,  -- Achievement
-    569593,  -- Level Up
-    568016,  -- Reputation (8473)
-    567400,  -- Quest Accepted
-    567439   -- Quest Turned In
-}
-
--- Vanilla
-muteSoundIDs_v = {
-    569593,  -- Level Up
-    568016,  -- Reputation (8473)
-    567400,  -- Quest Accepted
-    567439   -- Quest Turned In
+-- sounds.lua
+local soundConfig = {
+    retail = {
+        569143,  -- Achievement
+        1489546, -- Honor
+        569593,  -- Level Up
+        642841,  -- Battle Pet Level
+        4745441, -- Renown
+        568016,  -- Reputation
+        567400,  -- Quest Accepted
+        567439,  -- Quest Turned In
+        2066672  -- Trade Post
+    },
+    cata = {
+        569143,  -- Achievement
+        569593,  -- Level Up
+        568016,  -- Reputation
+        567400,  -- Quest Accepted
+        567439   -- Quest Turned In
+    },
+    vanilla = {
+        569593,  -- Level Up
+        568016,  -- Reputation
+        567400,  -- Quest Accepted
+        567439   -- Quest Turned In
+    }
 }
 
 --=====================================================================================
 -- Sound Options
 --=====================================================================================
 
-soundOptions = {
+local soundOptions = {
     "[Default]",
     "[Random]",
     "Altered Beast",
@@ -99,7 +97,7 @@ soundOptions = {
 --=====================================================================================
 -- Default Sounds
 --=====================================================================================
-defaultSounds = {
+local defaultSounds = {
     [1] = { -- Achievement
         [1] = "Interface\\Addons\\BLU\\sounds\\achievement_default_low.ogg",
         [2] = "Interface\\Addons\\BLU\\sounds\\achievement_default_med.ogg",
@@ -150,8 +148,7 @@ defaultSounds = {
 --=====================================================================================
 -- Custom Sounds
 --=====================================================================================
-
-sounds = {
+local sounds = {
     [3] = {
         [1] = "Interface\\Addons\\BLU\\sounds\\altered_beast_low.ogg",
         [2] = "Interface\\Addons\\BLU\\sounds\\altered_beast_med.ogg",